# src/main/resources/schema/swagger/wearenotnumbers-swagger-spec.yml
swagger: '2.0'
info:
  version: 1.0.0
  title:  We are not numbers
  description: ''
host: 'localhost:8080'
basePath: /
paths:
  '/v1/case':
      post:
        operationId: addCase
        summary: Add cases
        tags:
        - Case
        produces:
        - application/json
        parameters:
          - name: data
            in: body
            schema:
              $ref: '#/definitions/request-save-case'
        responses:
          '200':
            description: ''
            schema:
              $ref: '#/definitions/response-base'
          '500':
            $ref: '#/responses/response-internal-server-error-500'
  '/v1/case/{id}':
      get:
        operationId: getCaseById
        summary: Get single case by id
        tags:
        - Case
        produces:
        - application/json
        parameters: 
          - $ref: '#/parameters/id'
        responses:
          '200':
            description: ''
            schema:
              $ref: '#/definitions/response-case-single' 
          '500':
            $ref: '#/responses/response-internal-server-error-500'
  '/v1/case/all':
      get:
        operationId: getAllCases
        summary: Get All cases
        tags:
        - Case
        produces:
        - application/json
        parameters: 
          - name: page
            in: query
            type: integer
            format: int32
        responses:
          '200':
            description: ''
            schema:
              $ref: '#/definitions/response-case-list' 
          '500':
            $ref: '#/responses/response-internal-server-error-500'
parameters:
  id:
    in: path
    name: id
    type: integer
    format: int32
    required: true
definitions:
    response-base:
      type: object
      properties:
        success:
          type: boolean
        resultCode:
          type: integer
          format: int32
        message:
          type: string
        errors:
          type: array
          items:
            type: string
        transactionId:
          type: string
        type:
          type: string
      discriminator: type
      required:
        - success
        - resultCode
    model-case:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        history:
          type: string
        age:
          type: integer
          format: int32
        occupation:
          type: string
        city:
           type: string
      required:
        - name
        - email
        - history
    model-attribute:
      type: object
      properties:
        attName:
          type: string
        attType:
          type: string
        attCode:
          type: string
        attLabel:
          type: string
        attDesc:
          type: string
    model-enum-id-value:
      type: object
      properties:
        id: 
          type: integer
          format: int32
        value:
          type: string
      required:
         - id
    model-case-list:
      type: object
      properties:
        cases:
          type: array
          items: 
            $ref: '#/definitions/model-case'
    model-attribute-list:
      type: object
      properties:
        attributes:
          type: array
          items: 
            $ref: '#/definitions/model-attribute' 
    request-save-case:
      allOf:
        - $ref: '#/definitions/model-case'
    response-case-single:
      allOf:
        - $ref: '#/definitions/response-base'
        - type: object
          properties:
            returnValue:
              $ref: '#/definitions/model-case'
    response-case-list:
      allOf:
        - $ref: '#/definitions/response-base'
        - type: object
          properties:
            returnValue:
              $ref: '#/definitions/model-case-list'
responses:
    response-internal-server-error-500:
      description: Internal Server error
      schema:
        $ref: '#/definitions/response-base'
    response-unauthorized:
        description: Authentication information is missing or invalid
        schema:
          $ref: '#/definitions/response-base' 
        headers:
          WWW_Authenticate:
            type: string